Eagerly initialize SqsInstrumentationModule.
This is required to avoid enabled=false in some cases.
